requirements.txt
background
dash
dash_bootstrap_components
dash_bootstrap_templates
dash_extensions
geopy
pandas
plotly
pymongo
Requests
sqlalchemy
psycopg2-binary
dotenv